readprofile: make --mapfile opt argument optional
THe manpage for readprofile(8) states that we default
to '/boot/System.map' for the --mapfile option argument.
However, we define its optarg as required, which makes the
default value unusable. Let us make the optarg optional and
prevent error cases when the user only specifies --mapfile
without an argument on the command-line.

Signed-off-by: Christian Goeschel Ndjomouo <cgoesc2@wgu.edu>